+++ src/test/java/kitchenpos/application/TableServiceTest.java
@@ -1,9 +1,8 @@
 package kitchenpos.application;
 
 import kitchenpos.dao.OrderDao;
-import kitchenpos.dao.TableGroupDao;
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderStatus;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.domain.TableGroupRepository;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.domain.TableGroup;
